✨ feat: categoryValueToKey 함수 분리
