+++ src/api/meeting.ts
@@ -1,5 +1,4 @@
 import type {
-  // CreateMeetingRequest,
   CreateMeetingResponse,
   getMyMeetingsRequest,
   MeetingStatus,
