+++ src/pages/CreatePostPage.tsx
@@ -11,23 +11,7 @@ import {
   getOneYearLaterDateTime,
 } from '../utils/getLocalDateTime';
 import LoadingSpinner from '../components/LoadingSpinner';
-
-enum CategoryList {
-  KOREAN = '한식',
-  WESTERN = '양식',
-  CHINESE = '중식',
-  JAPANESE = '일식',
-  DESSERT = '디저트',
-  OTHER = '기타',
-}
-
-type CategoryKey = keyof typeof CategoryList;
-
-const categoryValueToKey = Object.keys(CategoryList).reduce((acc, key) => {
-  const value = CategoryList[key as CategoryKey];
-  acc[value] = key as CategoryKey;
-  return acc;
-}, {} as { [key: string]: CategoryKey });
+import { categoryValueToKey } from '../utils/categoryValueToKey';
 
 const CreatePostPage = () => {
   const {
